chore: rename `index` -> `async` file
